Ensure DPV output values are the same format as the machine generates.
Note: Trailing zeros are absent from the output when the machine saves a decimal value.

+++ src/main/groovy/com/seriouslypro/pnpconvert/PickSettings.groovy
@@ -17,14 +17,14 @@ class PickSettings {
     int placeSpeedPercentage = 100
     BigDecimal placeDelay = 0              // how long to wait after extending the pick nozzle when PLACING component, 0-5 seconds, resolution of 0.01 seconds
 
-    int takeHeight = 0              // ??
+    BigDecimal takeHeight = 0              // ??
 
     BigDecimal takeDelay = 0        // how long to wait after extending the pick nozzle when PICKING component, 0-3 seconds, resolution of 0.01 seconds.
                                     // using short takeDelay of around 0.25 seconds can prevent small components bouncing out of the tape.
 
-    int packageAngle = 0            // degrees, >= 0 < 360 - CLOCKWISE, relative to feeder angle
+    BigDecimal packageAngle = 0            // degrees, >= 0 < 360 - CLOCKWISE, relative to feeder angle
 
     // tape settings
-    int tapeSpacing = 4             // mm
+    BigDecimal tapeSpacing = 4             // mm
     int pullSpeed = 0
 }
